+++ indispensable-josef/src/jvmTest/kotlin/at/asitplus/signum/indispensable/josef/JwkTest.kt
@@ -144,7 +144,7 @@ class JwkTest : FreeSpec({
     }
 
     "Regression test: JWK (no keyId) -> CryptoPublicKey -> JWK (no keyId)" {
-        val key = randomCertificate().publicKey.toJsonWebKey()
+        val key = randomCertificate().decodedPublicKey.getOrThrow().toJsonWebKey()
         key.keyId shouldBe null
         val cpk = key.toCryptoPublicKey().getOrThrow()
         cpk.toJsonWebKey().keyId shouldBe null