+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/X509SignatureAlgorithm.kt
@@ -1,130 +1,184 @@
 package at.asitplus.signum.indispensable
 
 import at.asitplus.catching
+import at.asitplus.signum.UnsupportedCryptoException
 import at.asitplus.signum.indispensable.asn1.*
 import at.asitplus.signum.indispensable.asn1.encoding.Asn1
 import at.asitplus.signum.indispensable.asn1.encoding.Asn1.ExplicitlyTagged
 import at.asitplus.signum.indispensable.asn1.encoding.Asn1.Null
 import at.asitplus.signum.indispensable.asn1.encoding.decodeToInt
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.contract
 
-@Serializable(with = X509SignatureAlgorithmSerializer::class)
-enum class X509SignatureAlgorithm(
-    override val oid: ObjectIdentifier,
-    val isEc: Boolean = false
-) : Asn1Encodable<Asn1Sequence>, Identifiable, SpecializedSignatureAlgorithm {
+// future: SPI
+private interface X509SignatureAlgorithmProvider {
+    fun loaderForOid(oid: ObjectIdentifier): ((Asn1Structure.Iterator) -> X509SignatureAlgorithm)?
+}
+
+sealed class X509SignatureAlgorithmDescription(
+    override val oid: ObjectIdentifier
+) : Asn1Encodable<Asn1Sequence>, Identifiable {
+
+    /** Additional algorithm parameters. **Note: `null` means no parameters and `null != Asn1Null`!** */
+    abstract val parameters: Asn1Element?
+
+    override fun encodeToTlv() = Asn1.Sequence {
+        +oid
+        parameters?.let { +it }
+    }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is X509SignatureAlgorithmDescription) return false
+        return (oid == other.oid) && (parameters == other.parameters)
+    }
+
+    override fun hashCode() = (31 * oid.hashCode() + parameters.hashCode())
+
+    internal class Unknown(oid: ObjectIdentifier, override val parameters: Asn1Element?) :
+        X509SignatureAlgorithmDescription(oid)
+
+    companion object : Asn1Decodable<Asn1Sequence, X509SignatureAlgorithmDescription> {
+        override fun doDecode(src: Asn1Sequence) = src.decodeRethrowing {
+            val oid = next().asPrimitive().readOid()
+            // future: SPI
+            sequenceOf<X509SignatureAlgorithmProvider>(X509SignatureAlgorithm.Provider)
+                .firstNotNullOfOrNull { it.loaderForOid(oid) }
+                ?.invoke(this@decodeRethrowing)
+                ?: Unknown(oid, nextOrNull())
+        }
+    }
+}
+
+/** smart-casts the receiver to an [X509SignatureAlgorithm.Supported] if supported.*/
+@OptIn(ExperimentalContracts::class)
+fun X509SignatureAlgorithmDescription.isSupported(): Boolean {
+    contract {
+        returns(true) implies (this@isSupported is X509SignatureAlgorithm)
+    }
+    return (this is X509SignatureAlgorithm)
+}
+
+/** throws if the [X509SignatureAlgorithm] is unsupported */
+@OptIn(ExperimentalContracts::class)
+fun X509SignatureAlgorithmDescription.requireSupported() {
+    contract {
+        returns() implies (this@requireSupported is X509SignatureAlgorithm)
+    }
+    if (this !is X509SignatureAlgorithm) throw UnsupportedCryptoException("Unsupported X.509 signature algorithm (OID = ${this.oid})")
+}
+
+// future: open
+sealed class X509SignatureAlgorithm(
+    oid: ObjectIdentifier
+) : X509SignatureAlgorithmDescription(oid), SpecializedSignatureAlgorithm {
+
+    /** The [X509SignatureAlgorithmProvider] for Signum's natively supported [X509SignatureAlgorithm]s */
+    internal object Provider : X509SignatureAlgorithmProvider {
+        override fun loaderForOid(oid: ObjectIdentifier) = when (oid) {
+            KnownOIDs.rsaPSS -> X509SignatureAlgorithm::parsePssParams
+            else -> when (val alg = fromOid(oid)) {
+                null -> null
+                is RSAPKCS1 -> ({
+                    if (it.next() != Asn1Null) {
+                        throw Asn1TagMismatchException(
+                            Asn1Element.Tag.NULL, it.currentElement.tag,
+                            "RSA Params not allowed."
+                        )
+                    }
+                    alg
+                })
+
+                else -> ({ alg })
+            }
+        }
+    }
 
     // ECDSA with SHA-size
-    ES256(KnownOIDs.ecdsaWithSHA256, true),
-    ES384(KnownOIDs.ecdsaWithSHA384, true),
-    ES512(KnownOIDs.ecdsaWithSHA512, true),
+    sealed class ECDSA(oid: ObjectIdentifier, override val algorithm: SignatureAlgorithm.ECDSA) :
+        X509SignatureAlgorithm(oid) {
+        override val parameters get() = null
+        override val digest: Digest get() = algorithm.digest!!
+    }
+
+    @Deprecated("Use type check", replaceWith = ReplaceWith("this is X509SignatureAlgorithm.ECDSA"))
+    val isEc get() = this is ECDSA
 
     // RSASSA-PSS with SHA-size
-    PS256(KnownOIDs.rsaPSS),
-    PS384(KnownOIDs.rsaPSS),
-    PS512(KnownOIDs.rsaPSS),
+    sealed class RSAPSS(override val algorithm: SignatureAlgorithm.RSA) : X509SignatureAlgorithm(KnownOIDs.rsaPSS) {
+        override val digest get() = algorithm.digest
 
-    // RSASSA-PKCS1-v1_5 with SHA-size
-    RS256(KnownOIDs.sha256WithRSAEncryption),
-    RS384(KnownOIDs.sha384WithRSAEncryption),
-    RS512(KnownOIDs.sha512WithRSAEncryption),
-
-    // RSASSA-PKCS1-v1_5 using SHA-1
-    RS1(KnownOIDs.sha1WithRSAEncryption);
-
-    private fun encodePSSParams(bits: Int): Asn1Sequence =
-        when (bits) {
-            256 -> KnownOIDs.sha_256
-            384 -> KnownOIDs.sha_384
-            512 -> KnownOIDs.sha_512
-            else -> TODO()
-        }.let { shaOid ->
+        override val parameters by lazy {
+            val shaOid = digest.oid
+            val shaLength = digest.outputLength
             Asn1.Sequence {
-                +oid
-                +Asn1.Sequence {
-                    +ExplicitlyTagged(0u) {
+                +ExplicitlyTagged(0u) {
+                    +Asn1.Sequence {
+                        +shaOid
+                        +Null()
+                    }
+                }
+                +ExplicitlyTagged(1u) {
+                    +Asn1.Sequence {
+                        +KnownOIDs.pkcs1_MGF
                         +Asn1.Sequence {
                             +shaOid
                             +Null()
                         }
                     }
-                    +ExplicitlyTagged(1u) {
-                        +Asn1.Sequence {
-                            +KnownOIDs.pkcs1_MGF
-                            +Asn1.Sequence {
-                                +shaOid
-                                +Null()
-                            }
-                        }
-                    }
-                    +ExplicitlyTagged(2u) {
-                        +Asn1.Int(bits / 8)
-                    }
+                }
+                +ExplicitlyTagged(2u) {
+                    +Asn1.Int(shaLength.bytes)
                 }
             }
-        }
-
-    override fun encodeToTlv() = when (this) {
-        ES256, ES384, ES512 -> Asn1.Sequence { +oid }
-
-        PS256 -> encodePSSParams(256)
-
-        PS384 -> encodePSSParams(384)
-
-        PS512 -> encodePSSParams(512)
-
-        RS256, RS384, RS512, RS1 -> Asn1.Sequence {
-            +oid
-            +Null()
         }
     }
 
-    val digest: Digest
-        get() = when (this) {
-            RS1 -> Digest.SHA1
-            ES256, PS256, RS256 -> Digest.SHA256
-            ES384, PS384, RS384 -> Digest.SHA384
-            ES512, PS512, RS512 -> Digest.SHA512
-        }
+    // RSASSA-PKCS1-v1_5 with SHA-size
+    sealed class RSAPKCS1(oid: ObjectIdentifier, override val algorithm: SignatureAlgorithm.RSA) :
+        X509SignatureAlgorithm(oid) {
+        override val digest get() = algorithm.digest
+        override val parameters get() = Asn1Null
+    }
 
-    override val algorithm: SignatureAlgorithm
-        get() = when (this) {
-            ES256, ES384, ES512 -> SignatureAlgorithm.ECDSA(this.digest, null)
-            PS256, PS384, PS512 -> SignatureAlgorithm.RSA(this.digest, RSAPadding.PSS)
-            RS1, RS256, RS384, RS512 -> SignatureAlgorithm.RSA(this.digest, RSAPadding.PKCS1)
-        }
+    abstract val digest: Digest
+
+    object ES256 : ECDSA(KnownOIDs.ecdsaWithSHA256, SignatureAlgorithm.ECDSAwithSHA256)
+    object ES384 : ECDSA(KnownOIDs.ecdsaWithSHA384, SignatureAlgorithm.ECDSAwithSHA384)
+    object ES512 : ECDSA(KnownOIDs.ecdsaWithSHA512, SignatureAlgorithm.ECDSAwithSHA512)
+
+    object PS256 : RSAPSS(SignatureAlgorithm.RSAwithSHA256andPSSPadding)
+    object PS384 : RSAPSS(SignatureAlgorithm.RSAwithSHA384andPSSPadding)
+    object PS512 : RSAPSS(SignatureAlgorithm.RSAwithSHA512andPSSPadding)
+
+    object RS1 : RSAPKCS1(KnownOIDs.sha1WithRSAEncryption, SignatureAlgorithm.RSA(Digest.SHA1, RSAPadding.PKCS1))
+    object RS256 : RSAPKCS1(KnownOIDs.sha256WithRSAEncryption, SignatureAlgorithm.RSAwithSHA256andPKCS1Padding)
+    object RS384 : RSAPKCS1(KnownOIDs.sha384WithRSAEncryption, SignatureAlgorithm.RSAwithSHA384andPKCS1Padding)
+    object RS512 : RSAPKCS1(KnownOIDs.sha512WithRSAEncryption, SignatureAlgorithm.RSAwithSHA512andPKCS1Padding)
 
     companion object : Asn1Decodable<Asn1Sequence, X509SignatureAlgorithm> {
 
-        @Throws(Asn1OidException::class)
-        private fun fromOid(oid: ObjectIdentifier) = catching { entries.first { it.oid == oid } }.getOrElse {
-            throw Asn1OidException("Unsupported OID: $oid", oid)
+        //make it lazy to break init cycle that causes the weirdest nullpointer ever
+        val entries by lazy {
+            setOf(
+                ES256, ES384, ES512,
+                PS256, PS384, PS512,
+                RS1, RS256, RS384, RS512
+            )
         }
 
+        @Suppress("NOTHING_TO_INLINE")
+        private inline fun fromOid(oid: ObjectIdentifier) = entries.firstOrNull { it.oid == oid }
+
         @Throws(Asn1Exception::class)
-        override fun doDecode(src: Asn1Sequence): X509SignatureAlgorithm = src.decodeRethrowing {
-            when (val oid = next().asPrimitive().readOid()) {
-                ES512.oid, ES384.oid, ES256.oid -> fromOid(oid)
-
-                RS1.oid, RS256.oid, RS384.oid, RS512.oid -> fromOid(oid).also {
-                    val tag = next().tag
-                    if (tag != Asn1Element.Tag.NULL)
-                        throw Asn1TagMismatchException(Asn1Element.Tag.NULL, tag, "RSA Params not allowed.")
-                }
-
-                PS256.oid, PS384.oid, PS512.oid -> parsePssParams(this)
-                else -> throw Asn1Exception("Unsupported algorithm oid: $oid")
+        override fun doDecode(src: Asn1Sequence): X509SignatureAlgorithm =
+            X509SignatureAlgorithmDescription.doDecode(src).let {
+                (it as? X509SignatureAlgorithm)
+                    ?: throw Asn1OidException("Unsupported OID: ${it.oid}", it.oid)
             }
-        }
 
         @Throws(Asn1Exception::class)
-        private fun parsePssParams(src: Asn1Structure.Iterator): X509SignatureAlgorithm = runRethrowing{
+        private fun parsePssParams(src: Asn1Structure.Iterator): X509SignatureAlgorithm = runRethrowing {
             val (algSequence, mgfSequence, saltLen) = src.next().asSequence().decodeRethrowing {
                 Triple(
                     next().asExplicitlyTagged().verifyTag(0u).single().asSequence(),
@@ -196,18 +250,3 @@ fun SignatureAlgorithm.toX509SignatureAlgorithm() = catching {
 /** Finds a X.509 signature algorithm matching this algorithm. Curve restrictions are not preserved. */
 fun SpecializedSignatureAlgorithm.toX509SignatureAlgorithm() =
     this.algorithm.toX509SignatureAlgorithm()
-
-object X509SignatureAlgorithmSerializer : KSerializer<X509SignatureAlgorithm> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("X509SignatureAlgorithmSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: X509SignatureAlgorithm) {
-        value.let { encoder.encodeString(it.name) }
-    }
-
-    override fun deserialize(decoder: Decoder): X509SignatureAlgorithm {
-        val decoded = decoder.decodeString()
-        return X509SignatureAlgorithm.entries.first { it.name == decoded }
-    }
-}